requirements.txt
requests
pydantic
pydantic_settings
authlib
psutils